requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
alembic==1.17.0
    # via backend (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   starlette
azure-core==1.36.0
    # via
    #   azure-identity
    #   azure-keyvault-secrets
    #   azure-storage-blob
azure-identity==1.25.1
    # via backend (pyproject.toml)
azure-keyvault-secrets==4.10.0
    # via backend (pyproject.toml)
azure-storage-blob==12.27.0
    # via backend (pyproject.toml)
bcrypt==5.0.0
    # via backend (pyproject.toml)
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via uvicorn
cryptography==46.0.3
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
distro==1.9.0
    # via openai
ecdsa==0.19.1
    # via python-jose
exceptiongroup==1.3.0
    # via
    #   anyio
    #   pytest
fastapi==0.119.0
    # via backend (pyproject.toml)
flake8==7.3.0
    # via backend (pyproject.toml)
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   backend (pyproject.toml)
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
isodate==0.7.2
    # via
    #   azure-keyvault-secrets
    #   azure-storage-blob
jiter==0.11.1
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.0.0
    # via backend (pyproject.toml)
langchain-core==1.0.0
    # via
    #   langchain
    #   langchain-openai
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==1.0.0
    # via backend (pyproject.toml)
langgraph==1.0.0
    # via langchain
langgraph-checkpoint==2.1.2
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.0
    # via langgraph
langgraph-sdk==0.2.9
    # via langgraph
langsmith==0.4.37
    # via langchain-core
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via
    #   mako
    #   werkzeug
mccabe==0.7.0
    # via flake8
msal==1.34.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
openai==2.5.0
    # via
    #   backend (pyproject.toml)
    #   langchain-openai
orjson==3.11.3
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.11.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   pytest
passlib==1.7.4
    # via backend (pyproject.toml)
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.11
    # via backend (pyproject.toml)
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycodestyle==2.14.0
    # via flake8
pycparser==2.23
    # via cffi
pydantic==2.12.3
    # via
    #   backend (pyproject.toml)
    #   fastapi
    #   langchain
    #   langchain-core
    #   langgraph
    #   langsmith
    #   openai
pydantic-core==2.41.4
    # via pydantic
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via msal
pytest==8.4.2
    # via backend (pyproject.toml)
python-dotenv==1.1.1
    # via backend (pyproject.toml)
python-jose==3.5.0
    # via backend (pyproject.toml)
pyyaml==6.0.3
    # via langchain-core
regex==2025.9.18
    # via tiktoken
requests==2.32.5
    # via
    #   azure-core
    #   langsmith
    #   msal
    #   requests-toolbelt
    #   stripe
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.44
    # via
    #   backend (pyproject.toml)
    #   alembic
starlette==0.48.0
    # via fastapi
stripe==13.0.1
    # via backend (pyproject.toml)
tenacity==9.1.2
    # via langchain-core
tiktoken==0.12.0
    # via langchain-openai
tomli==2.3.0
    # via
    #   alembic
    #   pytest
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   azure-core
    #   azure-identity
    #   azure-keyvault-secrets
    #   azure-storage-blob
    #   cryptography
    #   exceptiongroup
    #   fastapi
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   stripe
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.37.0
    # via backend (pyproject.toml)
werkzeug==3.1.3
    # via backend (pyproject.toml)
xxhash==3.6.0
    # via langgraph
zstandard==0.25.0
    # via langsmith
